+++ reimi/frontend/reimi_app/lib/my_app.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:reimi_app/features/splash/presentation/splash_screen.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const SplashScreen(),
+    );
+  }
+}
+
